Enhance Versioner to use delayed evaluation of project version.
Using the Gradle project version as a delayed evaluation of
toString makes it possible to have Versioner process the options
without requiring preemptive creation of the verioner extension.

The GitData extension is also enabled as a lazy evaluation object
to ensure that we do not need to execute any logic until we have
to do it.

Additionally the @Memoized annotation reduces the lazy calls to
some internal methods to no more than once.