Avoid modifying already resolved configurations.
Other plugins or even build scripts may use single
file configurations and eagerly resolve them during
configuration stage. This avoids the problem that
would break the builds in Gradle 3+. Developers
may need to be given notice that they are resolving
the configurations eagerly and suggest using the
afterEvaluate approach. The order of plugin action
execution may need to be considered for these cases.